requirements.txt
cloudscraper>=1.2.66
google-auth>=2.14.1
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.1
google-cloud-logging>=3.2.5
google-cloud-storage>=2.6.0
google-resumable-media>=2.4.0